backend/LINEloginAPI/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Cors==4.0.0
psycopg2-binary==2.9.6
requests==2.31.0
python-dotenv==1.0.0
PyJWT==2.8.0